feat: refactor paywall export
