+++ typescript/packages/x402-hono/src/index.ts
@@ -6,10 +6,10 @@ import {
   computeRoutePatterns,
   findMatchingPaymentRequirements,
   findMatchingRoute,
-  getPaywallHtml,
   processPriceToAtomicAmount,
   toJsonSafe,
 } from "x402/shared";
+import { getPaywallHtml } from "x402/paywall";
 import {
   ERC20TokenAmount,
   FacilitatorConfig,
