+++ typescript/packages/x402/src/paywall/index.ts
@@ -1,4 +1,4 @@
-import { PAYWALL_TEMPLATE } from "../paywall/gen/template";
+import { PAYWALL_TEMPLATE } from "./gen/template";
 import { config } from "../types/shared/evm/config";
 import { PaymentRequirements } from "../types/verify";
