+++ typescript/packages/x402/src/shared/index.ts
@@ -1,5 +1,4 @@
 export * from "./json";
-export * from "./paywall";
 export * from "./base64";
 export * from "./network";
 export * from "./middleware";
